Run tests cross-platform
This should help to identify platform-specific failures.

Note that the test.sh script is broken down to its component parts which
are run individually in the Action. This was done to facilitate syntax
that would be compatible with Windows powershell, but is not extensible
and should be corrected if more test scenarios are added.